Testing Cookies and TimeAgo, Validation added

+++ src/main/resources/static/js/board.js
@@ -35,7 +35,9 @@ let index = {
     },
     delete: function () {
         let id = $("#id").val();
-
+        let r = confirm("이 글을 삭제하시겠습니까?");
+        if (r === false)
+            return;
         $.ajax({
             type: "DELETE",
             url: "/api/board/" + id,
@@ -53,6 +55,9 @@ let index = {
             title:$("#title").val(),
             content:$("#content").val()
         };
+        let r = confirm("글을 수정하시겠습니까?");
+        if (r === false)
+            return;
         $.ajax({
             type: "PUT",
             url: "/api/board/" + id + "/reply",
@@ -78,18 +83,8 @@ let index = {
             rootId: 0,
         };
 
-        $.ajax({
-            type: "POST",
-            url: `/api/login`,
-            data: JSON.stringify({username: data.userId, password: data.password}),
-            contentType: "application/json; charset=utf-8",
-            data_type: "json",
-            success: function(respData, status, xhr) {
-                window.token = xhr.getResponseHeader("Authentication");
-                console.log(window.token);
-            }
-            // 입력한 로그인 정보로 ajax 로그인 요청
-        }).done(function() {
+        this.login(data.userId, data.password)
+            .done(function() {
             console.log("댓글 쓰기 요청됨")
             $.ajax({
                 type: "POST",
@@ -110,7 +105,7 @@ let index = {
     },
     saveNestedReply: function (rootId) {
         let data = {
-            userId: $("#nestedReplyId-" + rootId).val(),
+            userId: $("#nestedReplyUsername-" + rootId).val(),
             password: $("#nestedReplyPassword-" + rootId).val(),
             boardId: $("#boardId").val(),
             content: $("#nestedReplyContent-" + rootId).val(),
@@ -119,22 +114,15 @@ let index = {
             rootId: rootId,
         };
 
-        console.log(data);
-
-        $.ajax({
-            type: "POST",
-            url: `/api/login`, // TODO: join & login을 하도록 바꾸기
-            data: JSON.stringify({username: data.userId, password: data.password}),
-            contentType: "application/json; charset=utf-8",
-            data_type: "json"
-            // 입력한 로그인 정보로 ajax 로그인 요청
-        }).done(function (resp) {
+        this.login(data.userId, data.password)
+            .done(function (resp) {
             $.ajax({
                 type: "POST",
                 url: `/api/board/${data.boardId}/reply`,
                 data: JSON.stringify(data),
                 contentType: "application/json; charset=utf-8",
-                data_type: "json"
+                data_type: "json",
+                headers: {"Authentication": 'Bearer ' + window.token},
             }).done(function (resp) {
                 alert("대댓글 쓰기가 완료되었습니다.");
                 location.href = `/board/${data.boardId}`;
@@ -146,10 +134,14 @@ let index = {
         });
     },
     replyDelete: function (boardId, replyId) {
+        let r = confirm("댓글을 삭제하시겠습니까?");
+        if (r === false)
+            return;
         $.ajax({
             type: "DELETE",
             url: `/api/board/${boardId}/${replyId}`,
-            data_type: "json"
+            data_type: "json",
+            headers: {"Authentication": 'Bearer ' + window.token},
         }).done(function (resp) {
             alert("댓글 삭제가 완료되었습니다.");
             location.href = `/board/${boardId}`;
@@ -157,7 +149,66 @@ let index = {
             alert(JSON.stringify(error));
         });
     },
+    modifyReplyTextareaToggle: function (replyId, selector) {
+        // let textarea = document.getElementById(selector + "TextareaId-" + replyId);
+        let textarea = $(`#${selector}TextareaId-${replyId}`)[0];
+        // let origContent = document.getElementById(selector + "ContentId-" + replyId);
+        let origContent = $(`#${selector}ContentId-${replyId}`)[0];
+        // let btnModify = document.getElementById(selector + "Modify-" + replyId);
+        let btnModify = $(`#${selector}Modify-${replyId}`)[0];
+        // let spanRegisterCancel = document.getElementById(selector + "OnModify-" + replyId);
+        let spanRegisterCancel = $(`#${selector}OnModify-${replyId}`)[0];
+
+        // if (document.getElementById("nestedReplyTextareaId-" + replyId).display === "none") {
+        if (textarea.style.display === "none") { // Textarea 열기
+            textarea.innerText = origContent.textContent;
+            textarea.style.display = "block";
+            origContent.style.display = "none";
+            btnModify.style.display = "none";
+            spanRegisterCancel.style.display = "inline-block";
+        } else { // Textarea 닫기
+            textarea.style.display = "none"
+            origContent.style.display = "block"
+            btnModify.style.display = "inline-block";
+            spanRegisterCancel.style.display = "none";
+        }
+
+    },
+    updateReply: function (boardId, replyId, selector) {
+        let r = confirm("댓글을 수정하시겠습니까?");
+        if (r === false)
+            return;
+        console.log("Authentication: Bearer " + window.token);
+        let content = {content: $(`#${selector}TextareaId-${replyId}`).val()};
+        $.ajax({
+            type: "PUT",
+            url: `/api/board/${boardId}/${replyId}`,
+            contentType: "application/json; charset=utf-8",
+            data: JSON.stringify(content),
+            headers: {"Authentication": 'Bearer ' + window.token},
+            xhrFields: {
+                withCredentials: true
+            },
+        }).done(function (resp) {
+            alert("대댓글 수정이 완료되었습니다.");
+            location.href = `/board/${boardId}`;
+        }).fail(function (error) {
+            alert(JSON.stringify(error));
+        });
+    },
+    login: function (userId, password) {
+        $.ajax({
+            type: "POST",
+            url: `/api/login`, // TODO: join & login을 하도록 바꾸기
+            data: JSON.stringify({username: userId, password: password}),
+            contentType: "application/json; charset=utf-8",
+            data_type: "json",
+            success: function(respData, status, xhr) {
+                window.token = xhr.getResponseHeader("Authentication");
+            }
+            // 입력한 로그인 정보로 ajax 로그인 요청
+        })
+    }
 
 };
-
 index.init();
